+++ libiview/parse_series_items.c
@@ -268,7 +268,7 @@ static void cdata_handler(void *_ctx, const xmlChar *data, int len) {
     }
 }
 
-ssize_t iv_parse_series_items(char *buf, size_t len, struct iv_item **items) {
+int iv_parse_series_items(char *buf, size_t len, struct iv_item **items) {
     // Instantiate SAX parser
     xmlSAXHandlerPtr handler = calloc(1, sizeof(xmlSAXHandler));
     handler->initialized = XML_SAX2_MAGIC;
